[1416] Move report views into page sub-routes
This makes these views use Wagtail's routing instead of having to
configure a separate URL in urls.py. This means that the privacy
settings of the report is checked and the URLs must be below the actual
URL of the report (you can't get to these views with an invalid program
now).