Exclude periods adjacent to HTML tags from endnote URLs
Sometimes endnotes contain URLs, punctuation, and HTML content that
run up against each other in a way resembling:

```
<p>text http://www.example.com.</p>
```

In our current regex-based URL detection scheme, the `.` character
would be included in the above URL.  We have a list of possible
character sequences that terminate a URL, and I've added `\.<` (a
literal period followed by an open pointy-bracket) to this list, which
lets us correctly exclude the period from the URL in situations like
the above.

It's possible that our list of exclusions will eventually get
unwieldy, but URL-finding in general is kind of hard and adding one
more right now seems like the least intrusive solution.