Forbid non-superusers from completing the copy action
This hook checks if the requesting user is a superuser, and if not,
rejects the copy.  The resulting "error message" (really just a 403
response) is not particularly friendly, but it's going to be mostly
impossible to reach this error message once the copy button is removed
for these users as well.