Remove "Copy" option from page listing dropdown buttons
This applies only if the user is not a superuser.  If they're not, we
drill down into the set of page listing buttons, find the dropdown
buttonsm, then remove the one labeled "Copy".  It's possible there are
other ways to do this, but this one seems straightforward enough and
uses a wagtail-provided hook.