Update wagtail from 2.8 to 2.10
In 2.9, `SiteMiddleware` was deprecated.  We weren't using
`request.site` anywhere, so I removed the middleware itself as not
needed.

In 2.10:

1. `get_document_model` import path changed
2. `<div class="rich-text">` wrappers removed from rich text.  Because
we are using `.rich-text` selectors in our CSS, I've chosen to
maintain the old behavior by adding a line to `INSTALLED_APPS`.  In
the future we may want to change this, but I think that would require
a more extensive review of rich text rendering across all the site
content than I am capable of giving right now.

See this link for more information:

https://docs.wagtail.io/en/stable/reference/contrib/legacy_richtext.html#legacy-richtext