Sync media files to local system by folder
This change simplifies and expands the implementation of the tasks:

* pull-{production,staging}-media
* pull-{production,staging}-images

Now, all four of these tasks use the same underlying
`pull_media_from_s3` function with different folders as arguments.  It
also changes how the `media` task works.  Instead of syncing all media
by default, it only syncs the `original_images` and `documents`
folders, which is usually what we want.  If that behavior needs to
change, the `folders` argument can be updated fairly easily.