Improve reliability on sometimes-flickering test
I've seen this test fail sometimes on CI and I'm never sure exactly
why, this is my attempt to get more information out of the assertion
failure.

What I think was happening here is the original test was not actually
testing what we wanted it to be testing, namely that *all* the results
are mostly in order.  We were just testing that the first result was
the most recent (and this was not always the case).

What I've done here is to move the test into its own class and remove
some additional sources of randomness that were interfering with how
ES was ordering the results.  Namely, I've given all the tests exactly
the same text data, rather than random fake data.