Add view for rendering the page to make a headless preview request
We need this view to, essentially, render the front end of the site
but with a way of letting the underlying page-fetching logic know that
it should request a special "preview" URL instead of its "normal"
endpoint.  This way it will render the page normally but use the
preview data.