Implement preview behavior on the front-end for reports
This mostly consists of:

1. Looking for an element with id = `na-react__preview`, meaning the
page we're looking at intended to be a preview page.
2. Getting the URL param values for the content type and the token.
These are supplied by wagtail-headless-preview and should be used when
querying the backend for the preview data.
3. Adding a route to the Reports page that corresponds to the URL for
wagtail-headless-preview's configured URL.
4. Making the report component know that it's supposed to be
previewing now, and if so, initiate the fetch request to the preview
URL with the appropriate content type and token data.

Having done this, the Report page renders as normal, except using
different data.