On preview, load the route being previewed
This is kind of a hack, but it does seem to work.  I'm not sure if
there's a nicer way to do it, since the `/h_preview` view only knows
what is being previewed *after* it looks up the page by its token,
which happens after the front-end is already mostly loaded.