Separate effects for query and pagination filter changes
Previously, the same code was run when the search query was updated as
when the forward/back page buttons were clicked.  This separates that
into two effects: one that functions as before for the page buttons,
and one that is called when the query changes which now causes the
current page filter to be reset.