Add initial CSP configuration
This set of changes configures the CSP based on my initial clicking
around the website locally, looking for problems.  Inevitably, there
will be things I have missed (due to lack of exhaustive searching and
differences between my environment and production).  Because of this,
I have set the CSP middleware to send "report-only" headers, meaning
the actual functioning of the site will not be affected yet.  It will
report violations to a URI that we configure to receive that
information, which we can then use to determine when we are ready to
activate the CSP for real.

I've set the CSP_REPORT_URI to send it to Sentry.io's service for
receiving these reports in the Heroku settings.