Move fellowship year calculation from model to form widget
Note: we have to add `typed_choice_field` to the classes so that
wagtail will display a nice big chevron on the select box.  This is an
internal wagtail CSS class that gets applied naturally to model fields
with choices.  Here, we are faking that, so we have to apply the class
in this way, so that the display is consistent with other choice
fields in the wagtail admin.