docx parsing: Build nested block elements in-place
This change attempts to reduce the amount of state tracked by the
`Block` class.  Rather than separately storing the current nested
element and appending it to the main list of elements only when we're
sure we can't append anything further to that nested element, keep
everything in the main list of elements all the time.

By keeping everything in one list, we ensure all sub-elements always
there.  Prior to this change, there were some cases when nested
elements were not being appended to the `_elements` attribute, such as
when a nested element was the last element in a document.

With this change, I've altered this class to not store the current
nested element, but to append it to `_elements` and to _check the end
of that list of elements_ for a potential nested element when adding a
new element.  This way it's impossible for a nested element to not be
included in the final list.

Fixes #1677