Make "data project external script" behave list "dataviz src"
Reports with a defined "data project external script" field load this
script by manipulating the DOM to add a `<script>` tag from within the
JS of the report component.  Reports with a defined "dataviz src"
field simply render the correct script tag from within the HTML
template.

This change makes the former behave more like the latter in terms of
the `<script>` tag.