Display featured articles on the Thread page
This component is a bit of a kludge but I'm not super sure how else to
do this efficiently with the existing `Fetch` and `ArticleListing`
components.  All `Featured` does is update the state of its parent to
include info about the Featured Articles so it can interleave them
with the existing "regular" articles.  I tried figuring out how to
make the `Fetch` component that owns the `ArticleListing` make two
requests, and it didn't really seem appropriate.  Also tried nesting
fetches, but that also seemed kind of gnarly and ultimately more
complicated than this solution.  So this is what I'm going with.