Enable optional text input fields
Previously, text input components were always required.  I've made
it possible to override this behavior by passing the prop `required`
as `false` if desired.

I have also added a red asterisk to all required fields, which is
still the default mode.

I have also slightly updated some of the CSS that controls the
animation.  I think the intent of the styles in `_inputs.scss` was to
make it so unselected/filled-in text inputs are kind of visually
closed and they "open" when you focus them, or when they have text
inside.  This effect was achieved, in part, using the `:valid`
pseudoclass, which applies to all empty text fields if we assume that
all text fields are required, which until this change was the case.
I've changed it so that now only required _and_ valid text fields are
given the "open" visual state by default, which I think preserves the
old behavior.