Use mailing list segment IDs instead of titles for signing up
Previously, the front end only tracked the "title" field for
displaying mailing list information.  But the title field has a
different value if a "display name" is defined for that mailing list,
and the front end doesn't know the difference between a real
title (Which is needed to actually tell Campaign Monitor what lists to
use), and a display name.

This commit solves the problem by adding a mailing list ID to the
information sent to the front end, which the subscribe component now
uses as the "source of truth" for which lists the user has checked,
and then sends back the list of IDs to the server, which the server
can use to find the true name of the list for use by campaign monitor.