reports: Handle links to other domains in Featured Section urls
Before this, if a featured section in a report was configured with a
URL that pointed to an outside domain (one different from
`location.origin`), the link would fail when clicked because it uses
the `Link` component from react router, which we've set up to work
only with relative link paths.

I've changed it so that we check the link and if it's to another
domain (using the pre-existing logic for checking that), and if so, we
render a normal `<a>` tag that will not interact with react router at
all.  If it _is_ a relative link, then it will use the react router
`Link` component as before.