Change featured link logic to handle internal, non-report links
If people use the featured section links to link to a different part
of the site, especially one that's not handled by react router, such
as a PDF file, then under the old logic that link won't work.  This
change allows that to work by expanding the set of links we use a
plain old `<a>` tag for to include any links outside of the report
currently being viewed.