Update program serializers to allow ordering of featured pages
This updates the program detail serializer (and the corresponding
subprogram serializer) to look in the GET request for a field named
`featured_pages_ordering`.  This can be either `asc` or `desc` and
changes the ordering of those pages that are stored on the
`story_grid` field to use that ordering.  The default mode is
descending.