Add fields to featured sub/program pages to control card display
Here I'm adding 3 fields to the Featured Program/Subprogram Pages to
more tightly control what's displayed in those cards.

1. Explicit title and story excerpt fields.  We use these when
serializing the objects if they exist, otherwise defer to the same
fields on the page being featured.
2. Boolean `show_page_type` field that is True by default, but can be
unchecked to hide the label of what type of page the page being
featured is.  Showing/hiding the label is done in the react component.